refactor: split out initializers
