+++ src/components/VtkTwoView.vue
@@ -189,12 +189,9 @@ import {
   toRefs,
   watch,
   watchEffect,
-  MaybeRef,
-  unref,
 } from 'vue';
 import { vec3 } from 'gl-matrix';
 import { onKeyStroke } from '@vueuse/core';
-import { Maybe } from '@/src/types';
 import { useResizeToFit } from '@/src/composables/useResizeToFit';
 import vtkLPSView2DProxy from '@/src/vtk/LPSView2DProxy';
 import vtkIJKSliceRepresentationProxy from '@/src/vtk/IJKSliceRepresentationProxy';
@@ -208,13 +205,14 @@ import BoundingRectangle from '@/src/components/tools/BoundingRectangle.vue';
 import { useToolSelectionStore } from '@/src/store/tools/toolSelection';
 import { useAnnotationToolStore } from '@/src/store/tools';
 import { doesToolFrameMatchViewAxis } from '@/src/composables/annotationTool';
-import type { TypedArray } from '@kitware/vtk.js/types';
 import { useSliceConfig } from '@/src/composables/useSliceConfig';
+import { useSliceConfigInitializer } from '@/src/composables/useSliceConfigInitializer';
 import { useWindowingConfig } from '@/src/composables/useWindowingConfig';
+import { useWindowingConfigInitializer } from '@/src/composables/useWindowingConfigInitializer';
 import { useResizeObserver } from '../composables/useResizeObserver';
 import { useOrientationLabels } from '../composables/useOrientationLabels';
 import { getLPSAxisFromDir } from '../utils/lps';
-import { useCurrentImage, useImage } from '../composables/useCurrentImage';
+import { useCurrentImage } from '../composables/useCurrentImage';
 import { useCameraOrientation } from '../composables/useCameraOrientation';
 import WindowLevelTool from './tools/windowing/WindowLevelTool.vue';
 import SliceScrollTool from './tools/SliceScrollTool.vue';
@@ -229,7 +227,6 @@ import { useDICOMStore } from '../store/datasets-dicom';
 import { useSegmentGroupStore } from '../store/segmentGroups';
 import vtkLabelMapSliceRepProxy from '../vtk/LabelMapSliceRepProxy';
 import { usePaintToolStore } from '../store/tools/paint';
-import useWindowingStore from '../store/view-configs/windowing';
 import { usePersistCameraConfig } from '../composables/usePersistCameraConfig';
 import CrosshairsTool from './tools/crosshairs/CrosshairsTool.vue';
 import { LPSAxisDir } from '../types/lps';
@@ -238,13 +235,7 @@ import { useViewProxy } from '../composables/useViewProxy';
 import { useWidgetManager } from '../composables/useWidgetManager';
 import useViewSliceStore from '../store/view-configs/slicing';
 import CropTool from './tools/crop/CropTool.vue';
-import {
-  ToolContainer,
-  VTKTwoViewWidgetManager,
-  WLAutoRanges,
-  WL_AUTO_DEFAULT,
-  WL_HIST_BINS,
-} from '../constants';
+import { ToolContainer, VTKTwoViewWidgetManager } from '../constants';
 import { useProxyManager } from '../composables/useProxyManager';
 import { getShiftedOpacityFromPreset } from '../utils/vtk-helpers';
 import { useLayersStore } from '../store/datasets-layers';
@@ -258,69 +249,6 @@ const SLICE_OFFSET_KEYS: Record<string, number> = {
   ArrowUp: -1,
   ArrowDown: 1,
 } as const;
-
-function useSliceConfigInitializer(
-  viewID: MaybeRef<string>,
-  imageID: MaybeRef<Maybe<string>>,
-  viewDirection: MaybeRef<LPSAxisDir>
-) {
-  const { metadata } = useImage(imageID);
-
-  const viewAxis = computed(() => getLPSAxisFromDir(unref(viewDirection)));
-  const sliceDomain = computed(() => {
-    const { lpsOrientation, dimensions } = metadata.value;
-    const ijkIndex = lpsOrientation[viewAxis.value];
-    const dimMax = dimensions[ijkIndex];
-
-    return {
-      min: 0,
-      max: dimMax - 1,
-    };
-  });
-
-  const store = useViewSliceStore();
-  const { config: sliceConfig } = useSliceConfig(viewID, imageID);
-
-  watch(
-    sliceConfig,
-    (config) => {
-      const imageIdVal = unref(imageID);
-      const viewIdVal = unref(viewID);
-      if (config || !imageIdVal) return;
-      store.updateConfig(viewIdVal, imageIdVal, {
-        ...sliceDomain.value,
-        axisDirection: unref(viewDirection),
-      });
-      store.resetSlice(viewIdVal, imageIdVal);
-    },
-    { immediate: true }
-  );
-}
-
-function useWindowingConfigInitializer(
-  viewID: MaybeRef<string>,
-  imageID: MaybeRef<Maybe<string>>
-) {
-  const { imageData } = useImage(imageID);
-
-  const store = useWindowingStore();
-  const { config: windowConfig } = useWindowingConfig(viewID, imageID);
-
-  watch(
-    windowConfig,
-    (config) => {
-      const image = imageData.value;
-      const imageIdVal = unref(imageID);
-      const viewIdVal = unref(viewID);
-      if (config || !image || !imageIdVal) return;
-
-      const [min, max] = image.getPointData().getScalars().getRange();
-      store.updateConfig(viewIdVal, imageIdVal, { min, max });
-      store.resetWindowLevel(viewIdVal, imageIdVal);
-    },
-    { immediate: true }
-  );
-}
 
 export default defineComponent({
   name: 'VtkTwoView',
@@ -355,7 +283,6 @@ export default defineComponent({
     SelectTool,
   },
   setup(props) {
-    const windowingStore = useWindowingStore();
     const viewSliceStore = useViewSliceStore();
     const viewCameraStore = useViewCameraStore();
     const layerColoringStore = useLayerColoringStore();
@@ -370,7 +297,6 @@ export default defineComponent({
     // --- computed vars --- //
 
     const {
-      currentImageData: curImageData,
       currentImageID: curImageID,
       currentImageMetadata: curImageMetadata,
       isImageLoading,
@@ -390,9 +316,6 @@ export default defineComponent({
       level: windowLevel,
     } = useWindowingConfig(viewID, curImageID);
 
-    const autoRange = computed<keyof typeof WLAutoRanges>(
-      () => wlConfig.value?.auto || WL_AUTO_DEFAULT
-    );
     const dicomInfo = computed(() => {
       if (
         curImageID.value != null &&
@@ -425,7 +348,18 @@ export default defineComponent({
 
     // --- initializers --- //
 
-    useSliceConfigInitializer(viewID, curImageID, viewDirection);
+    const sliceDomain = computed(() => {
+      const { lpsOrientation, dimensions } = curImageMetadata.value;
+      const ijkIndex = lpsOrientation[viewAxis.value];
+      const dimMax = dimensions[ijkIndex];
+
+      return {
+        min: 0,
+        max: dimMax - 1,
+      };
+    });
+
+    useSliceConfigInitializer(viewID, curImageID, viewDirection, sliceDomain);
     useWindowingConfigInitializer(viewID, curImageID);
 
     // --- view proxy setup --- //
@@ -482,111 +416,9 @@ export default defineComponent({
 
     // --- widget manager --- //
 
+    // TODO Attach to ViewProxy
     const { widgetManager } = useWidgetManager(viewProxy);
     provide(VTKTwoViewWidgetManager, widgetManager);
-
-    // --- window/level setup --- //
-
-    const histogram = (
-      data: number[] | TypedArray,
-      dataRange: number[],
-      numberOfBins: number
-    ) => {
-      const [min, max] = dataRange;
-      const width = (max - min + 1) / numberOfBins;
-      const hist = new Array(numberOfBins).fill(0);
-      data.forEach((value) => hist[Math.floor((value - min) / width)]++);
-      return hist;
-    };
-
-    const autoRangeValues = computed(() => {
-      // Pre-compute the auto-range values
-      if (curImageData?.value) {
-        const scalarData = curImageData.value.getPointData().getScalars();
-        const [min, max] = scalarData.getRange();
-        const hist = histogram(scalarData.getData(), [min, max], WL_HIST_BINS);
-        const cumm = hist.reduce((acc, val, idx) => {
-          const prev = idx !== 0 ? acc[idx - 1] : 0;
-          acc.push(val + prev);
-          return acc;
-        }, []);
-
-        const width = (max - min + 1) / WL_HIST_BINS;
-        return Object.fromEntries(
-          Object.entries(WLAutoRanges).map(([key, value]) => {
-            const startIdx = cumm.findIndex(
-              (v: number) => v >= value * 0.01 * scalarData.getData().length
-            );
-            const endIdx = cumm.findIndex(
-              (v: number) =>
-                v >= (1 - value * 0.01) * scalarData.getData().length
-            );
-            const start = Math.max(min, min + width * startIdx);
-            const end = Math.min(max, min + width * endIdx + width);
-            return [key, [start, end]];
-          })
-        );
-      }
-      return {};
-    });
-
-    const firstTag = computed(() => {
-      if (
-        curImageID.value &&
-        curImageID.value in dicomStore.imageIDToVolumeKey
-      ) {
-        const volKey = dicomStore.imageIDToVolumeKey[curImageID.value];
-        const { WindowWidth, WindowLevel } = dicomStore.volumeInfo[volKey];
-        return {
-          width: WindowWidth.split('\\')[0],
-          level: WindowLevel.split('\\')[0],
-        };
-      }
-      return {};
-    });
-
-    watch(
-      curImageData,
-      (imageData) => {
-        if (curImageID.value == null || wlConfig.value != null || !imageData) {
-          return;
-        }
-
-        const range = autoRangeValues.value[autoRange.value];
-        windowingStore.updateConfig(viewID.value, curImageID.value, {
-          min: range[0],
-          max: range[1],
-        });
-        if (firstTag.value?.width) {
-          windowingStore.updateConfig(viewID.value, curImageID.value, {
-            preset: {
-              width: parseFloat(firstTag.value.width),
-              level: parseFloat(firstTag.value.level),
-            },
-          });
-        }
-        windowingStore.resetWindowLevel(viewID.value, curImageID.value);
-      },
-      {
-        immediate: true,
-      }
-    );
-
-    watch(autoRange, (percentile) => {
-      if (
-        curImageID.value == null ||
-        wlConfig.value == null ||
-        !curImageData.value
-      ) {
-        return;
-      }
-      const range = autoRangeValues.value[percentile];
-      windowingStore.updateConfig(viewID.value, curImageID.value, {
-        min: range[0],
-        max: range[1],
-      });
-      windowingStore.resetWindowLevel(viewID.value, curImageID.value);
-    });
 
     // --- scene setup --- //
